get real pid start time
One cannot rely on the timestamps of files in /proc/<pid> : they might
be reset on suspend, or just a process getting SIGSTOP.

Use the real process start time instead: this requires reading
/proc/<pid>/stat, and converting from clocks to boot time to elapsed
time.

Use the nix crate here instead of libc crate because of the ergonomics;
and instead of uapi crate because it's faster to build thanks to having
less dependencies, and even faster once we only enable the features we
need.